Create phyiscal.txt as part of eml2eal
In order to create EML from web-hosted  entities, a mechanism is
required for providing the physical information required to fully
describe entities. This function does the reverse of that, creating
the physical.txt template from an EML file. This function will be
useful in an instance of updating a data package with large
entities as the editor will not need to download the large entities.