Fixes #110 (#115)
* Create phyiscal.txt as part of eml2eal

In order to create EML from web-hosted  entities, a mechanism is
required for providing the physical information required to fully
describe entities. This function does the reverse of that, creating
the physical.txt template from an EML file. This function will be
useful in an instance of updating a data package with large
entities as the editor will not need to download the large entities.

* Add entity URLs during eml2make_eml

* Fix the quote character issue

The utility function write_templates would not output
a table correctly due to the quote character. Manually
write to table using same functionality but with quote
parameter set to 'auto'.